+++ runtime/src/main/java/com/instana/android/fragments/FragmentHierarchyType.kt
@@ -1,0 +1,13 @@
+/*
+ * IBM Confidential
+ * PID 5737-N85, 5900-AG5
+ * Copyright IBM Corp. 2024, 2024
+ */
+
+package com.instana.android.fragments
+
+enum class FragmentHierarchyType {
+    PARALLEL,
+    INNER_CHILD,
+    SINGLE
+}